+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # v2.9.2
 
-* add extra metrics to `MetrifiedCacheDict`
-
+* refactored metrics in `MetrifiedCacheDict`